EasyModeViewControllerのintをNSIntegerに置き換えた

+++ DogCatch(ver1.01)/Controllers/TutorialScreenViewController.m
@@ -18,7 +18,7 @@
 
 @implementation TutorialScreenViewController
 {
-    int pageNumber;
+    NSInteger pageNumber;
 }
 
 - (void)viewDidLoad {
